+++ mw/index.ts
@@ -5,10 +5,11 @@ import {openPreference} from './preference';
 import {msg, setI18N, welcome, REPO_CDN, localize} from './msg';
 
 // 每次新增插件都需要修改这里
-const baseVersion = '3.1',
-	addons = ['escape'];
+const baseVersion = '3.3',
+	addons = ['theme'];
 
-mw.loader.load(`${CDN}/${REPO_CDN}/mediawiki.css`, 'text/css');
+// mw.loader.load(`${CDN}/${REPO_CDN}/mediawiki.css`, 'text/css');
+mw.loader.load('http://localhost:8080/codemirror-mediawiki/mediawiki.css', 'text/css');
 
 /**
  * jQuery.val overrides for CodeMirror.